refactor: move head and nav items to layout component
