+++ app/components/index.ts
@@ -1,0 +1,6 @@
+export { default as BCGovTypography } from './BCGovTypography';
+export { default as FooterLinks } from './FooterLinks';
+export { default as FormDiv } from './FormDiv';
+export { default as Layout } from './Layout';
+export { default as NavBarLinks } from './NavBarLinks';
+export { default as StyledGovButton } from './StyledGovButton';
